fix: cast head outputs to float32 and reduce margin to prevent underflow
Force score and classification head outputs to float32 to avoid bfloat16 underflow causing identical scores. Reduce margin ranking loss margin from 1.0 to 0.1 to help model differentiate scores more easily in lower precision. Add score statistics to training logs for monitoring.

+++ Model_Training/lens/model.py
@@ -125,8 +125,9 @@ class LENS(nn.Module):
         last_hidden_state = outputs.hidden_states[-1][torch.arange(batch_size, device=input_ids.device), sequence_lengths, :]
         
         # Pass through the two heads
-        score = self.score_head(last_hidden_state)
-        logits = self.classification_head(last_hidden_state)
+        # Force float32 for loss calculation to prevent underflow in bfloat16 causing identical scores
+        score = self.score_head(last_hidden_state).to(torch.float32)
+        logits = self.classification_head(last_hidden_state).to(torch.float32)
         
         return score, logits
